requirements.txt
bs4
selenium
webdriver-manager
discordwebhook
requests
python-dotenv
flask
pymongo
schedule